requirements.txt
decouple>=0.0.7
Django>=4.0,<5.0
djangorestframework>=3.14.0
dj-rest-auth>=2.1.9
django-allauth>=0.50.0
dj-database-url>=2.2.0
python-decouple>=3.6
django-cors-headers>=3.13.0
requests>=2.28.0
gunicorn>=20.1.0
PyJWT==2.8.0